requirements.txt
alembic==1.16.4
bcrypt==4.3.0
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysqlclient==2.2.7
pillow==11.2.1
psycopg2-binary==2.9.10
python-barcode==0.15.1
python-dateutil==2.9.0.post0
reportlab==4.4.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn